Finish when home

+++ myoro_player/lib/shared/blocs/user_preferences_cubit.dart
@@ -35,7 +35,7 @@ final class UserPreferencesCubit extends Cubit<UserPreferences> {
   Future<void> setVolume(double value) async {
     emit((await _userPreferencesService.update(
       data: {
-        UserPreferences.volumeJsonKey: value.toInt(),
+        UserPreferences.volumeJsonKey: value,
       },
     ))!);
   }
